Moves mixins to separate file; Removes Player as a parent class
